requirements.txt
langchain==0.0.208
deeplake
openai
tiktoken
deeplake
python-dotenv
streamlit